Refactor into first stable version. Added PGEN compatibility. Clean unused arguments. Changed defaults.
