--- neural-admixture/neural-admixture
@@ -1,7 +1,0 @@
-#!/usr/bin/env bash
-
-if [[ "$1" = -i ]]; then
-    python3 -m inference $@
-else
-    python3 -m train $@
-fi
