--- neural-admixture/src/parallel.py
@@ -1,8 +1,0 @@
-import torch.nn as nn
-
-class CustomDataParallel(nn.DataParallel):
-    def __getattr__(self, name):
-        try:
-            return super().__getattr__(name)
-        except AttributeError:
-            return getattr(self.module, name)
